Split sensor functions into separate files for each sensor
