+++ CMakeLists.txt
@@ -22,7 +22,9 @@ target_sources(app PRIVATE src/app_rpc.c)
 target_sources(app PRIVATE src/app_settings.c)
 target_sources(app PRIVATE src/app_state.c)
 target_sources(app PRIVATE src/app_sensors.c)
-target_sources(app PRIVATE src/sensors.c)
+target_sources(app PRIVATE src/sensor_bme280.c)
+target_sources(app PRIVATE src/sensor_scd4x.c)
+target_sources(app PRIVATE src/sensor_sps30.c)
 target_sources(app PRIVATE external/sensirion/embedded-common/common/sensirion_common.c)
 target_sources(app PRIVATE external/sensirion/embedded-common/i2c/sensirion_i2c_hal.c)
 target_sources(app PRIVATE external/sensirion/embedded-common/i2c/sensirion_i2c.c)
