+++ src/app_sensors.h
@@ -16,6 +16,7 @@
 
 #include <golioth/client.h>
 
+void app_sensors_init(void);
 void app_sensors_set_client(struct golioth_client *sensors_client);
 void app_sensors_read_and_stream(void);
